pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "LuscasV",email = "luscas.pessoal19@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-extensions (>=4.1,<5.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
factory-boy = "^3.3.3"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bookstore.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]